requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.7.3=py37h27cfd23_1
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h7f8727e_0
async-timeout=3.0.1=py37_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2022.3.29=h06a4308_0
cachetools=4.2.0=pyhd3eb1b0_0
certifi=2021.10.8=py37h06a4308_2
cffi=1.14.4=py37h261ae71_0
chardet=3.0.4=py37h06a4308_1003
click=8.1.2=pypi_0
cryptacular=1.5.5=pypi_0
cryptography=2.9.2=py37h1ba5d50_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
cython=0.29.28=py37h295c915_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
easydict=1.9=pypi_0
entrypoints=0.3=pypi_0
faiss-cpu=1.6.5=pypi_0
filelock=3.0.12=pypi_0
flask=2.1.1=pypi_0
flask-cors=3.0.10=pypi_0
freetype=2.10.4=h5ab3b9f_0
gdown=3.12.2=pypi_0
google-auth=1.24.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
grpcio=1.31.0=py37hf8bcb03_0
hupper=1.10.2=pypi_0
idna=2.10=py_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.21.0=py37h888b3d9_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
itsdangerous=2.1.2=pypi_0
jedi=0.18.0=py37h06a4308_1
jinja2=3.0.2=pypi_0
joblib=1.0.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=4.1.2=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=py37h06a4308_0
markupsafe=2.0.1=py37h27cfd23_0
matplotlib=3.3.3=pypi_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=4.7.6=py37h7b6447c_1
nbclient=0.5.4=pypi_0
nbconvert=6.2.0=pypi_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
ninja=1.10.2=py37hff7bd54_0
notebook=6.4.8=py37h06a4308_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1n=h7f8727e_0
packaging=21.0=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.1=pyhd3eb1b0_0
pastedeploy=2.1.1=pypi_0
pbkdf2=1.3=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py37he98fc37_0
pip=20.3.3=py37h06a4308_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
prettytable=2.0.0=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37he6710b0_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.8.0=pyhd3eb1b0_0
pyjwt=2.0.0=py37h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyramid=2.0=pypi_0
pyramid-mailer=0.15.1=pypi_0
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python3-openid=3.2.0=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda10.2.89_cudnn7.6.5_0
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=py37h2531618_1
readline=8.0=h7b6447c_0
repoze-sendmail=4.4.1=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7=pyhd3eb1b0_0
scikit-learn=0.24.0=pypi_0
scipy=1.6.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=51.1.2=py37h06a4308_3
six=1.15.0=py37h06a4308_0
sqlalchemy=1.3.23=pypi_0
sqlite=3.33.0=h62c20be_0
strongunreidusl=1.0.0=dev_0
tabulate=0.8.7=pypi_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=py37h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
torchvision=0.7.0=py37_cu102
tornado=6.1=py37h27cfd23_0
tqdm=4.56.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
transaction=3.0.1=pypi_0
translationstring=1.4=pypi_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd3eb1b0_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
webob=1.8.7=pypi_0
werkzeug=2.1.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h06a4308_0
wtforms=2.3.3=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yarl=1.5.1=py37h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zope-deprecation=4.4.0=pypi_0
zope-interface=5.2.0=pypi_0
zope-sqlalchemy=1.3=pypi_0
zstd=1.4.5=h9ceee32_0
